requirements.txt
asgiref==3.9.1
click==8.2.1
Django==5.2.5
django-redis==6.0.0
django-rq==3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
redis==6.4.0
rq==2.4.1
sqlparse==0.5.3
whitenoise==6.9.0
